Refactor junit output logic into pkg/compare/output.go
